split object tabs
