+++ frontend/app/src/entities/tasks/ui/task-tab.tsx
@@ -1,0 +1,28 @@
+import { TASK_TAB } from "@/config/constants";
+import { QSP } from "@/config/qsp";
+import { ObjectDetailsTab, TaskTabProps } from "@/entities/nodes/object/ui/object-tabs";
+import { useGetTaskCount } from "@/entities/tasks/domain/get-node-task-count/get-task-count.query";
+import { constructPath } from "@/shared/api/rest/fetch";
+import { Badge } from "@/shared/components/ui/badge";
+import { Spinner } from "@/shared/components/ui/spinner";
+import { useLocation } from "react-router";
+import { StringParam, useQueryParam } from "use-query-params";
+
+export function ObjectTaskTab({ objectId, ...props }: TaskTabProps) {
+  const { isPending, data: taskCount } = useGetTaskCount({ nodeId: objectId });
+  const [qspTab] = useQueryParam(QSP.TAB, StringParam);
+
+  const { pathname } = useLocation();
+
+  return (
+    <ObjectDetailsTab
+      isActive={qspTab === TASK_TAB}
+      to={constructPath(pathname, [{ name: QSP.TAB, value: TASK_TAB }])}
+      {...props}
+    >
+      Tasks
+      {isPending && <Spinner />}
+      <Badge className="font-medium rounded-full text-gray-80">{taskCount}</Badge>
+    </ObjectDetailsTab>
+  );
+}
